backend-image/requirements.txt
fastapi
pydantic
gunicorn
uvicorn
python-multipart
numpy
Pillow
torch
torchvision